requirements.txt
aenum==3.1.16
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.10.0
async-timeout==5.0.1
attrs==25.3.0
av==15.0.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
coloredlogs==15.0.1
dataclasses-json==0.6.7
deepgram-sdk==4.8.1
deprecation==2.1.0
distro==1.9.0
docstring_parser==0.17.0
eval_type_backport==0.2.2
exceptiongroup==1.3.0
filelock==3.19.1
Flask==3.1.2
flask-cors==6.0.1
flatbuffers==25.2.10
frozenlist==1.7.0
fsspec==2025.7.0
google-api-core==2.25.1
google-api-python-client==2.179.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
google-cloud-speech==2.33.0
google-cloud-texttospeech==2.27.0
google-genai==1.32.0
googleapis-common-protos==1.70.0
groq==0.31.0
grpcio==1.74.0
grpcio-status==1.74.0
gspread==6.2.1
h11==0.16.0
hf-xet==1.1.9
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.34.4
humanfriendly==10.0
idna==3.10
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
livekit==1.0.12
livekit-agents==1.2.7
livekit-api==1.0.5
livekit-blingfire==1.0.0
livekit-plugins-deepgram==1.2.7
livekit-plugins-google==1.2.7
livekit-plugins-groq==1.2.7
livekit-plugins-noise-cancellation==0.2.5
livekit-plugins-openai==1.2.7
livekit-plugins-silero==1.2.7
livekit-plugins-turn-detector==1.2.7
livekit-protocol==1.0.5
MarkupSafe==3.0.2
marshmallow==3.26.1
mpmath==1.3.0
multidict==6.6.4
mypy_extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.4.2
numpy==2.2.6
oauthlib==3.3.1
onnxruntime==1.22.1
openai==1.102.0
opentelemetry-api==1.36.0
opentelemetry-exporter-otlp==1.36.0
opentelemetry-exporter-otlp-proto-common==1.36.0
opentelemetry-exporter-otlp-proto-grpc==1.36.0
opentelemetry-exporter-otlp-proto-http==1.36.0
opentelemetry-proto==1.36.0
opentelemetry-sdk==1.36.0
opentelemetry-semantic-conventions==0.57b0
packaging==25.0
pillow==11.3.0
prometheus_client==0.22.1
propcache==0.3.2
proto-plus==1.26.1
protobuf==6.32.0
psutil==7.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.10.1
pyparsing==3.2.3
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
regex==2025.7.34
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
safetensors==0.6.2
sniffio==1.3.1
sounddevice==0.5.2
sympy==1.14.0
tenacity==9.1.2
tokenizers==0.21.4
torch==2.8.0
tqdm==4.67.1
transformers==4.55.4
types-protobuf==6.30.2.20250822
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.5.0
watchfiles==1.1.0
websockets==15.0.1
Werkzeug==3.1.3
yarl==1.20.1
zipp==3.23.0
